+++ src/http/temp/ASocket.hpp
@@ -55,7 +55,6 @@ class ASocket {
     socklen_t addrSize;
     socket_t server_socket;
     socket_t client_socket;
-    socket_t  sock;
 
 
     ASocket(void);
